Update copy used for Analytics conversion tracking.
