+++ assets/js/components/conversion-tracking/ConversionTrackingToggle.js
@@ -17,11 +17,7 @@
 /**
  * WordPress dependencies
  */
-import {
-	createInterpolateElement,
-	useState,
-	Fragment,
-} from '@wordpress/element';
+import { useState, Fragment } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 
 /**
@@ -31,7 +27,6 @@ import { Switch } from 'googlesitekit-components';
 import Data from 'googlesitekit-data';
 import { CORE_SITE } from '../../googlesitekit/datastore/site/constants';
 import ErrorText from '../../components/ErrorText';
-import Link from '../Link';
 import LoadingWrapper from '../LoadingWrapper';
 import ConfirmDisableConversionTrackingDialog from './ConfirmDisableConversionTrackingDialog';
 import { useFeature } from '../../hooks/useFeature';
@@ -39,19 +34,13 @@ import PropTypes from 'prop-types';
 
 const { useDispatch, useSelect } = Data;
 
-export default function ConversionTrackingToggle( { loading } ) {
+export default function ConversionTrackingToggle( { children, loading } ) {
 	const iceEnabled = useFeature( 'conversionInfra' );
 	const [ saveError ] = useState( null );
 	const [ showConfirmDialog, setShowConfirmDialog ] = useState( false );
 
 	const isConversionTrackingEnabled = useSelect( ( select ) =>
 		select( CORE_SITE ).isConversionTrackingEnabled()
-	);
-
-	const conversionTrackingDocumentationURL = useSelect( ( select ) =>
-		select( CORE_SITE ).getDocumentationLinkURL(
-			'enhanced-conversion-tracking'
-		)
 	);
 
 	const isSaving = useSelect( ( select ) =>
@@ -100,24 +89,7 @@ export default function ConversionTrackingToggle( { loading } ) {
 				tabletHeight="84px"
 			>
 				<p className="googlesitekit-settings-module__fields-group-helper-text">
-					{ createInterpolateElement(
-						__(
-							'Conversion tracking allows you to measure additional events on your site from other plugins that Site Kit integrates with to optimize your campaign performance. <a>Learn more</a>',
-							'google-site-kit'
-						),
-						{
-							a: (
-								<Link
-									href={ conversionTrackingDocumentationURL }
-									external
-									aria-label={ __(
-										'Learn more about conversion tracking',
-										'google-site-kit'
-									) }
-								/>
-							),
-						}
-					) }
+					{ children }
 				</p>
 			</LoadingWrapper>
 			{ showConfirmDialog && (
@@ -136,5 +108,6 @@ export default function ConversionTrackingToggle( { loading } ) {
 }
 
 ConversionTrackingToggle.propTypes = {
+	children: PropTypes.node.isRequired,
 	loading: PropTypes.bool,
 };
